requirements.txt
alembic==1.15.2
fastapi==0.116.1
jose==1.0.0
limits==5.5.0
pydantic==2.11.7
python-dotenv==1.1.1
Requests==2.32.5
slowapi
SQLAlchemy==2.0.43
sqlalchemy_schemadisplay==2.0
starlette==0.47.3

